+++ test/test_cli.sh
@@ -229,7 +229,7 @@ new "cli success validate"
 expectpart "$($clixon_cli -1 -f $cfg -l o validate)" 0 "^$"
 
 new "cli compare diff"
-expectpart "$($clixon_cli -1 -f $cfg -l o show compare text)" 0 "+            address 1.2.3.4"
+expectpart "$($clixon_cli -1 -f $cfg -l o show compare text)" 0 "+         address 1.2.3.4"
 
 new "cli start shell"
 expectpart "$($clixon_cli -1 -f $cfg -l o shell echo foo)" 0 "foo"